+++ src/main/java/me/scyphers/customcraft/gui/AddCraftingRecipeGUI.java
@@ -1,20 +1,22 @@
 package me.scyphers.customcraft.gui;
 
-import me.scyphers.customcraft.CustomCraft;
-import me.scyphers.customcraft.ui.GUI;
 import me.scyphers.customcraft.ui.InventoryGUI;
 import me.scyphers.customcraft.ui.ItemBuilder;
+import me.scyphers.customcraft.ui.Session;
 import me.scyphers.customcraft.ui.StaticGUI;
 import org.bukkit.Material;
 import org.bukkit.NamespacedKey;
-import org.bukkit.entity.Player;
-import org.bukkit.event.inventory.ClickType;
 import org.bukkit.event.inventory.InventoryAction;
 import org.bukkit.event.inventory.InventoryClickEvent;
-import org.bukkit.inventory.*;
+import org.bukkit.inventory.ItemStack;
+import org.bukkit.inventory.Recipe;
+import org.bukkit.inventory.ShapedRecipe;
+import org.bukkit.inventory.ShapelessRecipe;
 import org.jetbrains.annotations.NotNull;
 
-import java.util.*;
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.Map;
 
 public class AddCraftingRecipeGUI extends InventoryGUI {
 
@@ -25,35 +27,13 @@ public class AddCraftingRecipeGUI extends InventoryGUI {
     private boolean confirm = false;
     private boolean shapeless = false;
 
-    private final ItemStack[] ingredients = new ItemStack[9];
-    private ItemStack result = new ItemStack(Material.AIR);
-
-    /*
-     * Crafting Interface
-     * #########
-     * #000#####
-     * #000#C#0#
-     * #000#####
-     * #########
-     * B###S###E
-     * 0 - player enabled slot
-     * C - button to add recipe
-     * B - back button
-     * S - toggle shapeless/shaped
-     * E - Exit
-     */
-
-    public AddCraftingRecipeGUI(@NotNull CustomCraft plugin, @NotNull Player player, UUID viewer, NamespacedKey key) {
-        super(plugin, player, viewer, "<dark_purple>Add Crafting Recipe</dark_purple>", 54);
+    public AddCraftingRecipeGUI(Session session, NamespacedKey key) {
+        super(session, "<dark_purple>Add Crafting Recipe</dark_purple>", 54);
         this.key = key;
     }
 
     @Override
     public void draw() {
-
-        System.out.println("ingredients: " + Arrays.toString(ingredients));
-
-        // Default appearance
         fill(BACKGROUND);
         setItem(45, new ItemBuilder(Material.COMPASS).name("<gold>Back</gold>").build());
         setItem(53, new ItemBuilder(Material.IRON_DOOR).name("<red>Exit</red>").build());
@@ -61,77 +41,43 @@ public class AddCraftingRecipeGUI extends InventoryGUI {
         // Generate the ingredient view
         for (int column = 0; column < 3; column++) {
             for (int row = 0; row < 3; row++) {
-                int index = getIngredientIndex(column, row);
                 int inventoryIndex = getInventoryIndex(column, row);
-                setItem(inventoryIndex, getItem(ingredients[index]));
+                setItem(inventoryIndex, AIR);
             }
         }
 
-        // Add the Crafting Table
-        String addItemText = confirm ? "<gray>Click again to confirm adding this item!</gray>" : "<gray>Click to add this item!</gray>";
-        setItem(23, new ItemBuilder(Material.CRAFTING_TABLE).name("<gold>Add Recipe!</gold>").lore(addItemText).build());
-
-        // Add the result slot
-        setItem(25, getItem(result));
-
-        // Add the shapeless/shaped toggle
-        String shapeName = shapeless ? "<gold>Shapeless</gold>" : "<gold>Shaped</gold>";
-        String lore = "<dark_gray>Click to toggle to change the recipe to a " + shapeName + " recipe</dark_gray>";
-        Material type = shapeless ? Material.WATER_BUCKET : Material.ICE;
-        setItem(49, new ItemBuilder(type).name(shapeName).lore("").lore(lore).build());
+        setItem(23, confirmButton());
+        setItem(25, AIR);
+        setItem(49, toggleShapeButton());
 
     }
 
     @Override
-    public @NotNull GUI<?> handleInteraction(InventoryClickEvent event) {
-
+    public InventoryGUI onClick(InventoryClickEvent event) {
         int slot = event.getRawSlot();
+
+        // Block illegal actions that can pull from the inventory
         if (!isValidAction(event.getAction())) {
             event.setCancelled(true);
-        }
-
-        // Player inventory click
-        if (slot >= this.getSize()) {
             return this;
         }
 
-        System.out.println("cursor: " + event.getCursor());
-        System.out.println("current: " + event.getCurrentItem());
-        System.out.println("action: " + event.getAction());
-        System.out.println("click type: " + event.getClick());
-
-        ItemStack cursor = getItem(event.getCursor()).clone();
-
-        // Result slot click
-        if (slot == 25) {
-            if (cursor.getType() != Material.AIR) {
-                result = cursor;
-                System.out.println("updating result item");
-            } else {
-                result = AIR;
-                System.out.println("removing result item");
-            }
+        // Players can always use their inventory
+        if (slot >= this.getSize()) {
             return this;
         }
 
         int column = slot % 9 - 1;
         int row = slot / 9 - 1;
 
-        // Crafting inventory click
-        if (column >= 0 && column <= 2 && row >= 0 && row <= 2) {
-            int index = column + row * 3;
-            if (cursor.getType() != Material.AIR) {
-                ingredients[index] = cursor;
-                System.out.println("Updating ingredient item");
-            } else {
-                ingredients[index] = AIR;
-                System.out.println("Removing ingredient item");
-            }
+        // Crafting inventory / result click
+        if (column >= 0 && column <= 2 && row >= 0 && row <= 2 || slot == 25) {
+            event.setCancelled(false);
             return this;
         }
 
+        // Other interaction - block it
         event.setCancelled(true);
-
         return switch (slot) {
 
             // Add recipe
@@ -139,35 +85,38 @@ public class AddCraftingRecipeGUI extends InventoryGUI {
 
                 if (!confirm) {
                     confirm = true;
-                    this.draw();
+                    this.setItem(23, confirmButton());
+                    this.setUpdate(true);
                     yield this;
                 }
 
-                if (!validRecipe()) {
-                    getPlugin().getMessenger().chat(getPlayer(), "crafting.invalidCraftingRecipe");
+                ItemStack[] ingredients = getIngredients();
+                ItemStack result = item(getItem(25));
+
+                if (!isValidRecipe(ingredients, result)) {
+                    getSession().chat("crafting.invalidCraftingRecipe");
                     confirm = false;
                     yield this;
                 }
 
                 Recipe recipe = shapeless ? createShapelessRecipe(ingredients, result) : createShapedRecipe(ingredients, result);
-                getPlugin().getServer().addRecipe(recipe);
-                getPlugin().getMessenger().chat(getPlayer(), "crafting.addedCraftingRecipe", "%key%", key.getKey());
-
-                this.setShouldClose(true);
+                getSession().getPlugin().getServer().addRecipe(recipe);
+                getSession().chat("crafting.addedCraftingRecipe", "%key%", key.getKey());
+                this.setClose(true);
                 yield new StaticGUI(this);
-
             }
 
             // Toggle Shaped/Shapeless
             case 49 -> {
                 this.shapeless = !this.shapeless;
+                this.setItem(49, toggleShapeButton());
                 yield this;
             }
 
             // Navigation
-            case 45 -> new CraftHomeGUI(getPlugin(), getPlayer(), getViewer(), key);
+            case 45 -> new CraftHomeGUI(getSession(), key);
             case 53 -> {
-                this.setShouldClose(true);
+                this.setClose(true);
                 yield new StaticGUI(this);
             }
 
@@ -177,9 +126,9 @@ public class AddCraftingRecipeGUI extends InventoryGUI {
 
     }
 
-    @Override
-    public boolean allowPlayerInventoryEdits() {
-        return true;
+    @NotNull
+    private ItemStack item(ItemStack item) {
+        return item != null && item.getType() != Material.AIR ? item : AIR;
     }
 
     private int getIngredientIndex(int column, int row) {
@@ -190,21 +139,20 @@ public class AddCraftingRecipeGUI extends InventoryGUI {
         return 10 + column + row * 9;
     }
 
-    @NotNull
-    private ItemStack getItem(ItemStack item) {
-        return item != null && item.getType() != Material.AIR ? item : AIR;
-    }
+    private ItemStack[] getIngredients() {
 
-    private boolean isPlayerEditableSlot(int slot) {
-        if (slot > this.getSize()) return true;
-        return switch (slot) {
-            case 10, 11, 12, 19, 20, 21, 28, 29, 30, 25 -> true;
-            default -> false;
-        };
-    }
+        ItemStack[] ingredients = new ItemStack[9];
+        ItemStack[] inventory = getInventory().getContents();
 
-    private boolean isPermittedPlayerEdit(ClickType type) {
-        return type == ClickType.RIGHT || type == ClickType.LEFT;
+        for (int column = 0; column < 3; column++) {
+            for (int row = 0; row < 3; row++) {
+                int ingredientIndex = column + row * 3;
+                int inventoryIndex = getInventoryIndex(column, row);
+                ingredients[ingredientIndex] = inventory[inventoryIndex];
+            }
+        }
+
+        return ingredients;
     }
 
     private boolean isValidAction(InventoryAction action) {
@@ -213,6 +161,11 @@ public class AddCraftingRecipeGUI extends InventoryGUI {
                     PLACE_ALL, PLACE_SOME, PLACE_ONE, SWAP_WITH_CURSOR -> true;
             default -> false;
         };
+    }
+
+    private boolean isValidRecipe(ItemStack[] ingredients, ItemStack result) {
+        if (item(result).getType() == Material.AIR) return false;
+        return !Arrays.stream(ingredients).allMatch(item -> item(item).getType() == Material.AIR);
     }
 
     private Recipe createShapedRecipe(ItemStack[] ingredients, ItemStack result) {
@@ -238,9 +191,16 @@ public class AddCraftingRecipeGUI extends InventoryGUI {
         return recipe;
     }
 
-    private boolean validRecipe() {
-        if (getItem(result).getType() == Material.AIR) return false;
-        return !Arrays.stream(ingredients).allMatch(item -> getItem(item).getType() == Material.AIR);
+    private ItemStack confirmButton() {
+        String addItemText = confirm ? "<gray>Click again to confirm adding this item!</gray>" : "<gray>Click to add this item!</gray>";
+        return new ItemBuilder(Material.CRAFTING_TABLE).name("<gold>Add Recipe!</gold>").lore(addItemText).build();
+    }
+
+    private ItemStack toggleShapeButton() {
+        String shapeName = shapeless ? "<gold>Shapeless</gold>" : "<gold>Shaped</gold>";
+        String lore = "<dark_gray>Click to toggle to change the recipe to a " + shapeName + " recipe</dark_gray>";
+        Material type = shapeless ? Material.WATER_BUCKET : Material.ICE;
+        return new ItemBuilder(type).name(shapeName).lore("").lore(lore).build();
     }
 
     private static record ParsedShape(String[] shape, Map<Character, ItemStack> ingredients) {
