+++ src/main/java/me/scyphers/customcraft/gui/CraftHomeGUI.java
@@ -1,23 +1,21 @@
 package me.scyphers.customcraft.gui;
 
-import me.scyphers.customcraft.CustomCraft;
-import me.scyphers.customcraft.ui.GUI;
-import me.scyphers.customcraft.ui.ItemBuilder;
+import me.scyphers.customcraft.ui.InventoryGUI;
 import me.scyphers.customcraft.ui.MenuGUI;
+import me.scyphers.customcraft.ui.Session;
+import me.scyphers.customcraft.ui.ItemBuilder;
 import org.bukkit.Material;
 import org.bukkit.NamespacedKey;
-import org.bukkit.entity.Player;
 import org.bukkit.event.inventory.ClickType;
+import org.bukkit.event.inventory.InventoryAction;
 import org.jetbrains.annotations.NotNull;
-
-import java.util.UUID;
 
 public class CraftHomeGUI extends MenuGUI {
 
     private final NamespacedKey key;
 
-    public CraftHomeGUI(@NotNull CustomCraft plugin, @NotNull Player player, UUID intendedViewer, NamespacedKey key) {
-        super(plugin, player, intendedViewer, "<dark_purple>Custom Crafting</dark_purple>", 27);
+    public CraftHomeGUI(@NotNull Session session, @NotNull NamespacedKey key) {
+        super(session, "<dark_purple>Custom Crafting</dark_purple>", 27);
         this.key = key;
     }
 
@@ -28,12 +26,12 @@ public class CraftHomeGUI extends MenuGUI {
     }
 
     @Override
-    public GUI<?> onClick(int slot, ClickType type) {
-        return slot == 13 ? new AddCraftingRecipeGUI(getPlugin(), getPlayer(), getViewer(), key) : this;
+    public InventoryGUI onClick(int slot, ClickType type, InventoryAction action) {
+        return slot == 13 ? new AddCraftingRecipeGUI(getSession(), key) : this;
     }
 
     @Override
-    public GUI<?> getPreviousGUI() {
+    public InventoryGUI getPreviousGUI() {
         return this;
     }
 }
